Remove Material Dialog dependency

+++ app/src/main/java/com/tortel/deploytrack/dialog/DeleteDialog.java
@@ -16,14 +16,14 @@
 package com.tortel.deploytrack.dialog;
 
 import android.app.Dialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.content.LocalBroadcastManager;
+import android.support.v7.app.AlertDialog;
 
-import com.afollestad.materialdialogs.DialogAction;
-import com.afollestad.materialdialogs.MaterialDialog;
 import com.google.firebase.analytics.FirebaseAnalytics;
 import com.tortel.deploytrack.Analytics;
 import com.tortel.deploytrack.Log;
@@ -59,34 +59,23 @@ public class DeleteDialog extends DialogFragment {
     @NonNull
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState) {
-        MaterialDialog.Builder builder = new MaterialDialog.Builder(getActivity());
-        builder.content(R.string.confirm, mName);
-        builder.title(R.string.delete);
-        builder.positiveText(R.string.delete);
-        builder.negativeText(R.string.cancel);
-        builder.onPositive(new MaterialDialog.SingleButtonCallback() {
-            @Override
-            public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
-                Log.v("Deleting " + mId);
-                //Delete it
-                DatabaseManager.getInstance(getActivity()).deleteDeployment(mId);
-                // Notify the app
-                Intent deleteIntent = new Intent(DatabaseManager.DATA_DELETED);
-                deleteIntent.putExtra(KEY_ID, mId);
-                LocalBroadcastManager.getInstance(getActivity()).sendBroadcast(deleteIntent);
-                // Log the event
-                FirebaseAnalytics.getInstance(getActivity())
-                        .logEvent(Analytics.EVENT_DELETED_DEPLOYMENT, null);
-            }
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        builder.setMessage(getString(R.string.confirm, mName));
+        builder.setTitle(R.string.delete);
+        builder.setPositiveButton(R.string.delete, (DialogInterface dialog, int which) -> {
+            Log.v("Deleting " + mId);
+            //Delete it
+            DatabaseManager.getInstance(getActivity()).deleteDeployment(mId);
+            // Notify the app
+            Intent deleteIntent = new Intent(DatabaseManager.DATA_DELETED);
+            deleteIntent.putExtra(KEY_ID, mId);
+            LocalBroadcastManager.getInstance(getActivity()).sendBroadcast(deleteIntent);
+            // Log the event
+            FirebaseAnalytics.getInstance(getActivity())
+                    .logEvent(Analytics.EVENT_DELETED_DEPLOYMENT, null);
         });
+        builder.setNegativeButton(R.string.cancel, null);
 
-        builder.onNegative(new MaterialDialog.SingleButtonCallback() {
-            @Override
-            public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
-                dismiss();
-            }
-        });
-
-        return builder.build();
+        return builder.create();
     }
 }
